+++ internal/cmd/errors.go
@@ -2,18 +2,31 @@ package cmd
 
 import (
 	"errors"
+	"fmt"
 
+	"github.com/dhth/punchout/internal/config"
 	"github.com/dhth/punchout/internal/mcp/tools"
 )
 
-func IsErrUnexpected(err error) bool {
+func HandleError(err error) (string, bool) {
+	var zero string
 	switch {
+	case errors.Is(err, config.ErrConfigFileNotFound):
+		return fmt.Sprintf(`
+Here's a sample config:
+
+---
+%s---
+
+New to punchout? Run 'punchout tour' for a quick introduction.
+`, config.SampleConfig), false
 	case errors.Is(err, tools.ErrCouldntAddToolToServer):
-		return true
+		return zero, true
 	case errors.Is(err, tools.ErrCouldntConstructInputSchema):
-		return true
+		return zero, true
 	case errors.Is(err, tools.ErrCouldntConstructOutputSchema):
-		return true
+		return zero, true
 	}
-	return false
+
+	return zero, false
 }