+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Built-in themes
 - Support TUI and MCP settings in `punchout.toml`
+- A guided introduction to punchout's main features, available via `punchout tour`
 
 ## [v1.4.0] - Mar 17, 2026